mejoras de codigo

+++ datagame.h
@@ -9,7 +9,7 @@ static sf::RenderWindow *currentWindow;
 static float getAngulo(sf::Vector2f);
 static float randomFloat(float LO, float HI);
 static sf::Vector2u diffTo(sf::Vector2u startPos, sf::Vector2u endPos);
-static sf::Vector2f aimDirNorm(sf::Vector2f Pos, sf::Vector2f target);
+static sf::Vector2f aimDirNorm(sf::Vector2f Pos, sf::Vector2i target);
 };
